+++ app_utils/allianceauth.py
@@ -1,16 +1,14 @@
-import re
 from functools import partial
-from typing import List
 
 from django.contrib.auth.models import Permission, User
 
 from allianceauth.notifications import notify
-from allianceauth.tests.auth_utils import AuthUtils
 from allianceauth.views import NightModeRedirectView
 
 from ._app_settings import APP_UTILS_NOTIFY_THROTTLED_TIMEOUT
 from .django import users_with_permission
 from .helpers import throttle
+from .testing import create_fake_user  # noqa: F401
 
 
 def notify_admins(message: str, title: str, level: str = "info") -> None:
@@ -97,42 +95,3 @@ def notify_throttled(
 def is_night_mode(request) -> bool:
     """Returns True if the current user session is in night mode, else False"""
     return NightModeRedirectView.night_mode_state(request)
-
-
-def create_fake_user(
-    character_id: int,
-    character_name: str,
-    corporation_id: int = None,
-    corporation_name: str = None,
-    corporation_ticker: str = None,
-    alliance_id: int = None,
-    alliance_name: str = None,
-    permissions: List[str] = None,
-) -> User:
-    """Create a fake user incl. main character and (optional) permissions.
-
-    Will use default corporation and alliance if not set.
-    """
-    username = re.sub(r"[^\w\d@\.\+-]", "_", character_name)
-    user = AuthUtils.create_user(username)
-    if not corporation_id:
-        corporation_id = 2001
-        corporation_name = "Wayne Technologies Inc."
-        corporation_ticker = "WTE"
-    if corporation_id == 2001:
-        alliance_id = 3001
-        alliance_name = "Wayne Enterprises"
-    AuthUtils.add_main_character_2(
-        user=user,
-        name=character_name,
-        character_id=character_id,
-        corp_id=corporation_id,
-        corp_name=corporation_name,
-        corp_ticker=corporation_ticker,
-        alliance_id=alliance_id,
-        alliance_name=alliance_name,
-    )
-    if permissions:
-        perm_objs = [AuthUtils.get_permission_by_name(perm) for perm in permissions]
-        user = AuthUtils.add_permissions_to_user(perms=perm_objs, user=user)
-    return user