+++ utils-test-app/utils_test_app/tests/test_allianceauth.py
@@ -16,41 +16,6 @@ from app_utils.helpers import throttle
 MODULE_PATH = "app_utils.allianceauth"
 
 
-class TestCreateFakeUser(TestCase):
-    def test_should_create_fake_user(self):
-        # when
-        user = create_fake_user(1001, "Bruce Wayne")
-        # then
-        self.assertTrue(User.objects.filter(pk=user.pk).exists())
-        self.assertEqual(user.username, "Bruce_Wayne")
-        self.assertEqual(user.profile.main_character.character_id, 1001)
-        self.assertEqual(user.profile.main_character.character_name, "Bruce Wayne")
-        self.assertEqual(user.profile.main_character.corporation_id, 2001)
-        self.assertEqual(user.profile.main_character.alliance_id, 3001)
-        self.assertEqual(user.profile.main_character.alliance_name, "Wayne Enterprises")
-
-    def test_should_create_fake_user_with_corporation(self):
-        # when
-        user = create_fake_user(
-            1001,
-            "Bruce Wayne",
-            corporation_id=2002,
-            corporation_name="Dummy corp",
-            corporation_ticker="ABC",
-        )
-        # then
-        self.assertEqual(user.profile.main_character.corporation_id, 2002)
-        self.assertEqual(user.profile.main_character.corporation_name, "Dummy corp")
-        self.assertEqual(user.profile.main_character.corporation_ticker, "ABC")
-        self.assertIsNone(user.profile.main_character.alliance_id)
-
-    def test_should_create_fake_user_with_permissions(self):
-        # when
-        user = create_fake_user(1001, "Bruce Wayne", permissions=["auth.add_group"])
-        # then
-        self.assertTrue(user.has_perm("auth.add_group"))
-
-
 class TestNotifyAdmins(TestCase):
     def test_should_notify_users_with_logging_permissions_and_superusers_only(self):
         # given