+++ utils-test-app/utils_test_app/tests/test_testing.py
@@ -7,6 +7,7 @@ from allianceauth.authentication.models import EveCharacter
 from app_utils.testing import (
     NoSocketsTestCase,
     SocketAccessError,
+    create_fake_user,
     create_user_from_evecharacter,
     generate_invalid_pk,
 )
@@ -58,3 +59,38 @@ class TestCreateUserFromEveCharacter(TestCase):
         self.assertEqual(character_ownership.character, self.character)
         self.assertEqual(character_ownership.user, user)
         self.assertTrue(user.token_set.filter(scopes__name="dummy_scope").exists())
+
+
+class TestCreateFakeUser(TestCase):
+    def test_should_create_fake_user(self):
+        # when
+        user = create_fake_user(1001, "Bruce Wayne")
+        # then
+        self.assertTrue(User.objects.filter(pk=user.pk).exists())
+        self.assertEqual(user.username, "Bruce_Wayne")
+        self.assertEqual(user.profile.main_character.character_id, 1001)
+        self.assertEqual(user.profile.main_character.character_name, "Bruce Wayne")
+        self.assertEqual(user.profile.main_character.corporation_id, 2001)
+        self.assertEqual(user.profile.main_character.alliance_id, 3001)
+        self.assertEqual(user.profile.main_character.alliance_name, "Wayne Enterprises")
+
+    def test_should_create_fake_user_with_corporation(self):
+        # when
+        user = create_fake_user(
+            1001,
+            "Bruce Wayne",
+            corporation_id=2002,
+            corporation_name="Dummy corp",
+            corporation_ticker="ABC",
+        )
+        # then
+        self.assertEqual(user.profile.main_character.corporation_id, 2002)
+        self.assertEqual(user.profile.main_character.corporation_name, "Dummy corp")
+        self.assertEqual(user.profile.main_character.corporation_ticker, "ABC")
+        self.assertIsNone(user.profile.main_character.alliance_id)
+
+    def test_should_create_fake_user_with_permissions(self):
+        # when
+        user = create_fake_user(1001, "Bruce Wayne", permissions=["auth.add_group"])
+        # then
+        self.assertTrue(user.has_perm("auth.add_group"))